requirements.txt
Flask
python-dotenv
requests
mysql-connector-python
google-generativeai
